Translations basically working
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS36]$/ResXResourceManager/Feature/Default;C34873

+++ ResXManager.Translate/ITranslationMatch.cs
@@ -12,7 +12,7 @@
             get;
         }
 
-        int Rating
+        double Rating
         {
             get;
         }
